youtube/requirements.txt
# YouTube 데이터 수집에 필요한 패키지

# YouTube Data API
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# YouTube 자막 추출
youtube-transcript-api==0.6.1

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0

# HTTP 요청
requests>=2.31.0

# 환경변수 관리
python-dotenv>=1.0.0

# 기타 유틸리티
tqdm>=4.66.0  # 진행률 표시
